// src/apis/models/FireBytes_Unified_PageFilters_Dtos_PageFilterDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { FireBytes_Unified_PageFilters_Dtos_PageFilterElementDto } from './FireBytes_Unified_PageFilters_Dtos_PageFilterElementDto';
export type FireBytes_Unified_PageFilters_Dtos_PageFilterDto = {
    id?: string | null;
    readonly extraProperties?: Record<string, any> | null;
    concurrencyStamp?: string | null;
    creationTime?: string;
    creatorId?: string | null;
    lastModificationTime?: string | null;
    lastModifierId?: string | null;
    pageFilterName?: string | null;
    pageFilterElements?: Array<FireBytes_Unified_PageFilters_Dtos_PageFilterElementDto> | null;
    layout?: string | null;
    labelAlign?: string | null;
    tenantId?: string | null;
};