// src/apis/models/FireBytes_Unified_PageFilters_Dtos_PageFilterElementDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type FireBytes_Unified_PageFilters_Dtos_PageFilterElementDto = {
    id?: string | null;
    readonly extraProperties?: Record<string, any> | null;
    concurrencyStamp?: string | null;
    creationTime?: string;
    creatorId?: string | null;
    lastModificationTime?: string | null;
    lastModifierId?: string | null;
    pageFilterId?: string | null;
    sortNum?: number | null;
    field?: string | null;
    filterTitle?: string | null;
    filterWidth?: number | null;
    queryType?: string | null;
    componentName?: string | null;
    skipValueIsNull?: boolean | null;
    args?: Record<string, any> | null;
    valueChange?: Array<string> | null;
    isAdvanced?: boolean | null;
    ifShow?: boolean | null;
    tenantId?: string | null;
};